проверка на наличие программы в датафайлах, если нет то создать копию

+++ IntegrityVerification/MyDataFile.cs
@@ -14,7 +14,8 @@ namespace IntegrityVerification
 
             if (File.Exists(Constants.dataFilePath) == false)
             {
-                File.Copy(Process.GetCurrentProcess().MainModule.FileName, Constants.appFilePath);
+                if (File.Exists(Constants.appFilePath) == false)
+                    File.Copy(Process.GetCurrentProcess().MainModule.FileName, Constants.appFilePath);
 
                 File.Create(Constants.dataFilePath).Close();
                 File.WriteAllText(Constants.dataFilePath, JsonConvert.SerializeObject(new List<TrackedObject>(), Formatting.Indented));
